geo-building-extractor/backend/extraction-service/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
GeoAlchemy2
requests
gunicorn
marshmallow
marshmallow-sqlalchemy